Added tests, cleaned code, and improved covering strategy
Abstract Covering Strategy now has two modes of patrolling non-covered vertices: deterministic and indeterministic. The anti-random and anti-waiting agent strategies use the deterministic mode and the anti-statistical strategy uses the indeterministic mode.

Removed the arrived method call for the agent strategy interface and replaced it with the thwarted method call. This way, the simulation can inform the agent when it has thwarted an attack, which can happen when the agent is occupying a vertex and an adversary chooses to attack at that time.